v3-dev Updates
These changes update the search API and structure of the v3 code to be more maintainable and support further features. At this point the search API supports basic searching and returns a valid set of OPDS `publications`. However these publications stil lack links and the `editions` extension.

Further work will involve ensuring that `search` links are available from all pages and adding an OPDS Publication class that will allow the code governing the creation of those objects to be reusable